+++ src/golang/lib/constants.go
@@ -1,3 +1,3 @@
-package constants
+package lib
 
 const ServerVersionNumber = "0.0.16"
