+++ src/golang/lib/job/k8s.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/aqueducthq/aqueduct/lib"
 	"github.com/aqueducthq/aqueduct/lib/collections/integration"
 	"github.com/aqueducthq/aqueduct/lib/collections/operator/function"
 	"github.com/aqueducthq/aqueduct/lib/collections/shared"
-	"github.com/aqueducthq/aqueduct/lib/constants"
 	"github.com/aqueducthq/aqueduct/lib/k8s"
 	"github.com/dropbox/godropbox/errors"
 	"k8s.io/client-go/kubernetes"
@@ -97,7 +97,7 @@ func (j *k8sJobManager) Launch(ctx context.Context, name string, spec Spec) erro
 	if err != nil {
 		return err
 	}
-	containerImage := fmt.Sprintf("%s:%s", containerRepo, constants.ServerVersionNumber)
+	containerImage := fmt.Sprintf("%s:%s", containerRepo, lib.ServerVersionNumber)
 
 	return k8s.LaunchJob(
 		name,
